==> multilayer_perceptron/4_pool_results_from_imputations_mlp.ipynb <==
# Pool results from the multilayer perceptron for all 5 sets of data
- Calculate a single predicted probability for each patient via averaging the score for each patient (usrds_id) across the restulst for the 5 (imputations) sets of the model results

import pickle
import numpy as np
import pandas as pd

import psycopg2
from sqlalchemy import create_engine

con = create_engine('postgresql://username:password@location/dbname')

### Import the imputed predicted probabilities

with open('./results/2021_MLP_final_ytest_all.pickle', 'rb') as f:  
    y_true_mlp = pickle.load(f)
    
with open('./results/2021_MLP_final_ypred_all.pickle', 'rb') as picklefile:  
    y_pred_proba_mlp = pickle.load(picklefile)

### Reformat into a pandas dataframe to make it easier to work with
The index for the values  is zero here b/c we only saved it for the positive class, otherwise it would be 1

pooled = pd.DataFrame()
pooled['imp1']=y_pred_proba_mlp[0][:,0] 
pooled['imp2']=y_pred_proba_mlp[1][:,0] 
pooled['imp3']=y_pred_proba_mlp[2][:,0] 
pooled['imp4']=y_pred_proba_mlp[3][:,0] 
pooled['imp5']=y_pred_proba_mlp[4][:,0] 
pooled.head()

### Calculate the mean and standard deviation of the predicted probability for the positive class (died_in_90) for each patient/row 

pooled['score'] = pooled.mean(axis=1)
pooled['std_'] = pooled.std(axis=1)
pooled.head()

### Import the details from the original data from step 1

dataset = pd.read_sql_query('''SELECT  usrds_id, died_in_90, subset FROM medxpreesrd;''', con)

### sort the values so they are in the same order as when the LR was calculated and keep only the test set

dataset = dataset[dataset.subset > 6].copy().sort_values(by = 'usrds_id').reset_index(drop=True)

### merge the details with the pooled predictions

pooled = pooled.merge(dataset, left_index=True, right_index=True)
pooled

### save as a pickle file

with open('./results/2021_final_MLP_model_test_pred_proba_pooled.pickle', 'wb') as picklefile:  
        pickle.dump(pooled, picklefile)
