==> multilayer_perceptron/6_mlp_fairness_assessment.ipynb <==
# Calculate a fairness assessment of the multilayer perceptron model

import numpy as np
import pandas as pd
import pickle

import sys
#path to the functions folder
sys.path.append('../onc_functions')

from fairness import get_fairness_assessment
print('python-' + sys.version)

import datetime
dte = datetime.datetime.now()
dte = dte.strftime("%Y")
print(dte)

import psycopg2
from sqlalchemy import create_engine

con = create_engine('postgresql://username:password@location/dbname')

## Import only the columns we want to look at in this assessment

df = pd.read_sql_query('''SELECT   usrds_id, died_in_90, inc_age, sex, dialtyp, race, hispanic, subset FROM medxpreesrd;''', con)

## Save these selected columns

with open('complete_fair1.pickle', 'wb') as picklefile:  
    pickle.dump(df, picklefile)

## Load fairness columns and model results

with open('complete_fair1.pickle', 'rb') as f:
        dataset = pickle.load(f)
dataset.head()

with open('./results/2021_final_MLP_model_test_pred_proba_pooled.pickle','rb') as f:  
    proba = pickle.load(f)
proba.head()

## Merge the results with the fairness details

data = dataset.merge(proba, on=['usrds_id','subset','died_in_90'])
data.head()

   # Fairness

fairness= get_fairness_assessment(data,
                                  y_proba_col_name='score',
                                  y_true_col_name='died_in_90')
fairness.head(39)

fairness.to_csv('./results/' + str(dte) + '_mlp_calibrated_fairness.csv')
