==> multilayer_perceptron/7_mlp_risk_assess.ipynb <==
# Calculate Risk categories from the multilayer perceptron model pooled results

!pip install --upgrade pip
!pip install pandas

import numpy as np
import pandas as pd
import pickle

import sys
#path to the functions folder
sys.path.append('../onc_functions')

from risk import get_risk_categories

print('python-' + sys.version)
import datetime
dte = datetime.datetime.now()
dte = dte.strftime("%Y%m%d")

# here we import the dataset we created for the fairness assessment in notebook 6
with open('complete_fair1.pickle','rb') as f:  
    details = pickle.load(f)
details.head()

with open('./results/2021_final_MLP_model_test_pred_proba_pooled.pickle','rb') as f:  
    y_pred = pickle.load(f)
y_pred.head()

data = details.merge(y_pred, on=['usrds_id','died_in_90','subset'])
data.head()

risk_cat = get_risk_categories(data,
                               y_proba_col_name='score',
                               y_true_col_name='died_in_90')
risk_cat

risk_cat.to_csv('./results/' + str(dte) + '_mlp_risk_cat.csv')
